Extract a reusable FieldAnnotator class
